Enable some Windows builds
Some of the packages here can be built on Windows. This is because they
are not dependent on `cuda-cudart` (currently not built on Windows). So
enable those packages to build and just skip the packages that lack
`cuda-cudart`. While this doesn't cover everything, it is a good
incremental improvement that adds some more packages to Windows.