Deselect `files` on platforms where it is empty
Some of the packages in this feedstock have nothing in the `files` field
to include during packaging. However when conda-build parses this field,
it thinks it treats it as `null` (likely due to how the YAML is parsed)
and errors when tryin to iterate over it. To fix this issue, just hide
the `files` key when no entries are actually added.